add all events with enums to all model controllers

+++ deployChaincodeToNetwork.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-DEPLOYMENT_PATH=/srv/docker/kt-tam-extra_hosts-3node-2channel/fabric-samples/3node2channel/deployment
+
+# TODO: clean up this file its a command mess
+
+DEPLOYMENT_PATH=/srv/docker/hyperledger-fabric-extra_hosts-5orgs/fabric-samples/5node2channel/deployment
 CHAINCODE_DEPLOYMENT_PATH=/src/github.com/hyperledger/fabric/peer
 ABSOLUTE_PATH=/opt/gopath/src/github.com/chaincode/chaincode-solidary-network-chaincode
 # CHAINCODE=chaincode-solidary-network-chaincode
@@ -14,14 +17,16 @@ PEER0_ORG2_ID=2
 PEER0_ORG3_ID=3
 PEER0_ORG4_ID=4
 PEER0_ORG5_ID=5
-PEER_CHAINCODE_PATH=/srv/docker/kt-tam-extra_hosts-3node-2channel/fabric-samples/chaincode
+PEER_CHAINCODE_PATH=/srv/docker/hyperledger-fabric-extra_hosts-5orgs/fabric-samples/chaincode
 # deploy install/upgrade chaincode
-CHANNEL=channelall
+CHANNEL_ALL=channelall
 CHAINCODE_NAME=sncc
 CHAINCODE_CONVECTOR=solidary-network-chaincode
 CHAINCODE_LANG=node
-POLICY_CHANNELALL="'Org1MSP.member', 'Org2MSP.member','Org3MSP.member','Org4MSP.member','Org5MSP.member'"
+POLICY_CHANNEL_ALL="'Org1MSP.member', 'Org2MSP.member','Org3MSP.member','Org4MSP.member','Org5MSP.member'"
+# POLICY_CHANNEL_ALL="'Org1MSP.peer', 'Org2MSP.peer','Org3MSP.peer','Org4MSP.peer','Org5MSP.peer'"
 LANG=node
+CA_FILE=/var/hyperledger/configs/ordererOrganizations/example.com/orderers/orderer1.example.com/msp/tlscacerts/tlsca.example.com-cert.pem
 # tar tgz to push
 TGZ_FILENAME=chaincode-${CHAINCODE_CONVECTOR}.tgz
 TGZ_PATH=chaincode-${CHAINCODE_CONVECTOR}
@@ -36,21 +41,30 @@ GOV_NAME="Big Government"
 press_any_key() {
   read -n 1 -s -r -p "Press any key to continue";printf "\\n"
   :;
-  # sleep 1
+  # SEC=5
+  # echo "sleeping for ${SEC}sec..."
+  # sleep ${SEC}
 }
 
-# STARTER version v0 (occurs on cc:start)
+# STARTER version 1.0 (occurs on cc:start)
 # following UPDATE versions (occure on cc:upgrade)
+# ibm auction network
+# VERSION="1.10"
+# production network
 VERSION="1.0"
+# 1 build with hurley, 0 only when we want to skip restart hurley network to build the chaincode, with 0 we dont reBuild chaincode, good for just deploy to networks
+BUILD_WITH_HURLEY=1
 
 echo "confirm deploy '${CHAINCODE_CONVECTOR} ${VERSION}'"
-read -n 1 -s -r -p "Press any key to continue";printf "\\n\\n"
+read -n 1 -s -r -p "Press any key to continue";printf "\\n"
 
-# always clean up chaincode path
-# rm ${TGZ_PATH} -r || true
-# cleanup, remove pack after push to peers
-# rm ${TGZ_FILENAME} -r || true
-rm ${CHAINCODE_NAME}.pak || true
+if [ ${BUILD_WITH_HURLEY} -eq 1 ]; then
+  # always clean up chaincode path
+  # rm ${TGZ_PATH} -r || true
+  # cleanup, remove pack after push to peers
+  # rm ${TGZ_FILENAME} -r || true
+  rm ${CHAINCODE_NAME}.pak || true
+fi
 
 # start
 # npm run cc:start -- ${CHAINCODE_CONVECTOR}
@@ -63,15 +77,18 @@ if [ "${VERSION}" == "1.0" ]; then
   BUILD_CC_ACTION="start"
   DEPLOY_CC_ACTION="instantiate"
   # require restart env to prevent "Error: could not assemble transaction, err proposal response was not successful, error code 500, msg chaincode with name 'solidary-network-chaincode' already exists"
-  # TODO: this extra step is really necessary: ~30sec
-  npm run env:restart
-  npx lerna run build --scope @solidary-network/common-cc --stream
-  npm run cc:${BUILD_CC_ACTION} -- ${CHAINCODE_CONVECTOR}
+  if [ ${BUILD_WITH_HURLEY} -eq 1 ]; then
+    npm run env:restart
+    npx lerna run build --scope @solidary-network/common-cc --stream
+    npm run cc:${BUILD_CC_ACTION} -- ${CHAINCODE_CONVECTOR}
+  fi
 else
   BUILD_CC_ACTION="upgrade"
   DEPLOY_CC_ACTION="upgrade"
-  npx lerna run build --scope @solidary-network/common-cc --stream
-  npm run cc:${BUILD_CC_ACTION} -- ${CHAINCODE_CONVECTOR} ${VERSION}  
+  if [ ${BUILD_WITH_HURLEY} -eq 1 ]; then
+    npx lerna run build --scope @solidary-network/common-cc --stream
+    npm run cc:${BUILD_CC_ACTION} -- ${CHAINCODE_CONVECTOR} ${VERSION}  
+  fi
 fi
 
 # pack and push
@@ -88,56 +105,32 @@ do
   IP=${PEERS[${IDX}]}
     # if is node 1
   if [ $peer -eq 1 ]; then
-    # copy tgz file to node
-#   scp ${TGZ_FILENAME} ${IP}:/tmp
-    # remove old package folder before extract new one
-#   ssh -t ${IP} rm /tmp/chaincode-${CHAINCODE_CONVECTOR} -r
-    # extract tgz to tmp file
-#   ssh -t ${IP} tar xvf /tmp/${TGZ_FILENAME} -C /tmp
-    # copy extracted package to peer ABSOLUTE_PATH to build pak (fabric-samples/chaincode)
-#   ssh -t ${IP} docker cp /tmp/chaincode-${CHAINCODE_CONVECTOR} peer0.org1.example.com:${ABSOLUTE_PATH}
-    # package chaincode to PEER_CHAINCODE_PATH
-#   echo "package ${CHAINCODE_NAME}.pak"
-#   press_any_key
-    # -s, --cc-package create CC deployment spec for owner endorsements instead of raw CC deployment spec
-    # -S, --sign, if creating CC deployment spec package for owner endorsements, also sign it with local MSP
-#   ssh -t ${IP} docker exec cli peer chaincode package -l node -S -s -n ${CHAINCODE_NAME} -p ${ABSOLUTE_PATH} -v ${VERSION} ${CHAINCODE_NAME}.pak
-    # copy to package local file system
-#   echo "copy ${CHAINCODE_NAME}.pak to local file system"
-#   scp ${IP}:${CHAINCODE_DEPLOYMENT_PATH}/${CHAINCODE_NAME}.pak .
-#   press_any_key
-
-    echo "bring peer0.org1.hurley.lab packaged ${CHAINCODE_CONVECTOR}.${VERSION} to local file system ${CHAINCODE_NAME}.pak"
-    docker cp peer0.org1.hurley.lab:/var/hyperledger/production/chaincodes/${CHAINCODE_CONVECTOR}.${VERSION} ${CHAINCODE_NAME}.pak
-    press_any_key
+    if [ ${BUILD_WITH_HURLEY} -eq 1 ]; then
+      echo "bring peer0.org1.hurley.lab packaged ${CHAINCODE_CONVECTOR}.${VERSION} to local file system ${CHAINCODE_NAME}.pak"
+      docker cp peer0.org1.hurley.lab:/var/hyperledger/production/chaincodes/${CHAINCODE_CONVECTOR}.${VERSION} ${CHAINCODE_NAME}.pak
+      press_any_key
+    fi
 
     # copy to package to orderer deployment path /used only was backup
     echo "copy ${CHAINCODE_NAME}.pak to orderer deployment path"
     press_any_key
     scp ${CHAINCODE_NAME}.pak ${ORDERER_IP}:${DEPLOYMENT_PATH}
-  # else
-    # moved to outside if, now copy to all nodes/peers
-    # copy to package to peer deployment path
-    # echo "copy ${CHAINCODE_NAME}.pak to peer0Org${peer} deployment path"
-    # press_any_key
-    # scp ${CHAINCODE_NAME}.pak ${IP}:${CHAINCODE_DEPLOYMENT_PATH}
   fi
 
   echo "copy ${CHAINCODE_NAME}.pak to peer0Org${peer} deployment path"
   press_any_key
   scp ${CHAINCODE_NAME}.pak ${IP}:${DEPLOYMENT_PATH}
 
-  # copy to dontainer
-# ssh ${IP} docker cp ${CHAINCODE_DEPLOYMENT_PATH}/${CHAINCODE_NAME}.pak cli:/opt/gopath/src${CHAINCODE_DEPLOYMENT_PATH}/${CHAINCODE_NAME}.pak
+  # copy to container
   ssh ${IP} docker cp ${DEPLOYMENT_PATH}/${CHAINCODE_NAME}.pak cli:/opt/gopath/${CHAINCODE_DEPLOYMENT_PATH}/${CHAINCODE_NAME}.pak
   # always install on all peers
   # install on peer0Org?
-  echo "install ${CHAINCODE_NAME} in peer0Org${peer}"
-  ssh ${IP} docker exec cli peer chaincode install ${CHAINCODE_NAME}.pak -v ${VERSION}
+  echo "install ${CHAINCODE_NAME} ${VERSION} in peer0Org${peer}"
+  ssh ${IP} docker exec -e CORE_PEER_ADDRESS=peer0.org${peer}.example.com:7051 cli peer chaincode install ${CHAINCODE_NAME}.pak -v ${VERSION}
+  press_any_key
   # install on peer1Org?
   echo "install ${CHAINCODE_NAME} ${VERSION} in peer1Org${peer}"
-  press_any_key
-  ssh ${IP} docker exec -e CORE_PEER_ADDRESS=peer1.org${peer}.example.com:7051 cli peer chaincode install ${CHAINCODE_NAME}.pak
+  ssh ${IP} docker exec -e CORE_PEER_ADDRESS=peer1.org${peer}.example.com:7051 cli peer chaincode install ${CHAINCODE_NAME}.pak -v ${VERSION}
   press_any_key
 done
 
@@ -146,25 +139,7 @@ done
 # instantiate|upgrade chaincode on peer0Org1 in channelall with policy....only after all nodes are registered, and with chaincode installed
 echo "${DEPLOY_CC_ACTION} ${CHAINCODE_NAME} ${VERSION} in peer0Org1"
 press_any_key
-# ssh -t ${PEER0_ORG1_IP} docker exec cli peer chaincode ${DEPLOY_CC_ACTION} -o orderer.example.com:7050 -l ${CHAINCODE_LANG} -C ${CHANNEL} -n ${CHAINCODE_NAME} ${ABSOLUTE_PATH} -v ${VERSION} -c "'{ \"Args\":[]}' -P \"OR(${POLICY_CHANNELALL})\""
-ssh -t ${PEER0_ORG1_IP} docker exec cli peer chaincode ${DEPLOY_CC_ACTION} -o orderer.example.com:7050 -l ${CHAINCODE_LANG} -C ${CHANNEL} -n ${CHAINCODE_CONVECTOR} ${ABSOLUTE_PATH} -v ${VERSION} -c "'{ \"Args\":[]}' -P \"OR(${POLICY_CHANNELALL})\""
-
-# is the same as above with ${DEPLOY_CC_ACTION}
-# TODO instantiate or upgrade or both
-# TODO upgrade after install ?
-# TODO WARN -v ${VERSION} is required here too: seems yes: "Must supply value for chaincode name, path and version parameters."
-# TODO Version of the chaincode specified in install/instantiate/upgrade commands
-# if [ "${PEER0_ORG1_IP}" != "1.0" ]; then
-#   echo "upgrade ${CHAINCODE_NAME} ${VERSION} in peer1Org${peer}"
-#   press_any_key
-#   ssh ${PEER0_ORG1_IP} docker exec cli peer chaincode upgrade -o orderer.example.com:7050 -l ${CHAINCODE_LANG} -C ${CHANNEL} -n ${CHAINCODE_NAME} -v ${VERSION} -c "'{ \"Args\":[]}' -P \"OR(${POLICY_CHANNELALL})\""
-# fi
-
-# TODO: participant_createWithParameters participant_create don't work via ssh
-echo "now don't forget to:"
-echo "1. invoke gov and johndoe, to be tested on sanity-check.sh";
-echo "2. run sanity-check.sh, and be patient, until whole network is quick and responsive";
-exit
+ssh -t ${PEER0_ORG1_IP} docker exec cli peer chaincode ${DEPLOY_CC_ACTION} -o orderer1.example.com:7050 --tls --cafile ${CA_FILE} -l ${CHAINCODE_LANG} -C ${CHANNEL_ALL} -n ${CHAINCODE_CONVECTOR} ${ABSOLUTE_PATH} -v ${VERSION} -c "'{ \"Args\":[]}' -P \"OR(${POLICY_CHANNEL_ALL})\""
 
 for peer in "${PEERS_ID[@]}"
 do
@@ -173,46 +148,30 @@ do
   IP=${PEERS[${IDX}]}
   # list instantiated chaincodes on channelall
   echo "list instantiate chaincodes in peer0Org${peer}"
+  ssh ${IP} "docker exec -e CORE_PEER_ADDRESS=peer0.org${peer}.example.com:7051 cli peer chaincode list -C channelall --instantiated"
   press_any_key
-  ssh ${IP} "docker exec cli peer chaincode list -C channelall --instantiated"
   echo "list instantiate chaincodes in peer1Org${peer}"
-  press_any_key
   ssh ${IP} "docker exec -e CORE_PEER_ADDRESS=peer1.org${peer}.example.com:7051 cli peer chaincode list -C channelall --instantiated"
-
-  # TODO: leave for seed
-  # invoke participant_createWithParameters
-  # if [ $peer -eq 1 ]; then
-  #   # ssh -t ${IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{ \"Args\" : [\"participant_createWithParameters\", \"${GOV_ID}\", \"${GOV_CODE}\", \"${GOV_NAME}\" ] }'"
-  #   ssh -t ${IP} docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '"{ \"Args\" : [\"participant_create\", \"{ \"id\": \"${GOV_ID}\", \"code\": \"${GOV_CODE}\", \"name\" :\"${GOV_NAME}\" }\" ] }"'
-  #   press_any_key
-  # fi
-  # invoke chaincode participant_get
-  # ssh ${IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{ \"Args\" : [\"participant_get\", \"${GOV_ID}\" ] }'"
-  # ssh ${IP} "docker exec -e CORE_PEER_ADDRESS=peer1.org${peer}.example.com:7051 cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{ \"Args\" : [\"participant_get\", \"${GOV_ID}\" ] }'"
-  # # invoke chaincode participant_getByCode
-  # ssh ${IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{ \"Args\" : [\"participant_getByCode\", \"${GOV_CODE}\" ] }'"
-  # ssh ${IP} "docker exec -e CORE_PEER_ADDRESS=peer1.org${peer}.example.com:7051 cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{ \"Args\" : [\"participant_getByCode\", \"${GOV_CODE}\" ] }'"
-  # press_any_key
+  press_any_key
 done
 
-# TODO work on seeds
-exit
+# install indexs and views if is intantiated chaincode
+if [ "${VERSION}" == "1.0" ]; then
+  # TODO: replace with stuff inside chaincode
+  echo "install couchdb views and indexes..."
+  ./couchdb/installNetwork.sh
 
-# invoke chaincode participant_create
-# TODO this create must be 
-# WORKS ON PEER not here: docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{"Args":["participant_create", "{\"id\":\"c8ca045c-9d1b-407f-b9ae-31711758f2d0\",\"code\":\"gov\",\"name\":\"Big Government\"}"]}'
-# ssh -t ${PEER0_ORG1_IP} docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c "'{ \"Args\": [\"participant_create\", { \"id\": \"c8ca045c-9d1b-407f-b9ae-31711758f2d0\", \"code\": \"gov\", \"name\": \"Big Government\" } ] }'"
-# ssh -t ${PEER0_ORG1_IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{ \"Args\": [\"participant_create\", { \"id\": \"c8ca045c-9d1b-407f-b9ae-31711758f2d0\", \"code\": \"gov\", \"name\": \"Big Government\" } ] }'"
-# invoke chaincode participant_get
-# OK USED ABOVE
-# ssh -t ${PEER0_ORG1_IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{ \"Args\" : [\"participant_get\", \"c8ca045c-9d1b-407f-b9ae-31711758f2d0\" ] }'"
-# invoke chaincode participant_getByCode
-# OK USED ABOVE
-# ssh -t ${PEER0_ORG1_IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{\"Args\":[\"participant_getByCode\", \"gov\"]}'"
-# invoke chaincode person_create
-# WORKS ON PEER not here: docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{"Args":["person_create", "{\"id\":\"4ea88521-031b-4279-9165-9c10e1839001\",\"firstname\":\"John\",\"lastname\":\"Doe\",\"beneficiaryNumber\":\"285191659\",\"birthDate\":\"61985472\",\"cardVersion\":\"006.007.23\",\"country\":\"PRT\",\"documentNumber\":\"09879462 0 ZZ3\",\"documentType\":\"Cartão De Cidadão\",\"emissionDate\":\"61985472\",\"emittingEntity\":\"República Portuguesa\",\"expirationDate\":\"61985472\",\"fatherFirstname\":\"Alberto\",\"fatherLastname\":\"De Andrade Monteiro\",\"fiscalNumber\":\"182692124\",\"gender\":\"M\",\"height\":\"1.81\",\"identityNumber\":\"098794620\",\"motherFirstname\":\"Maria Da Graça De Oliveira Mendes\",\"motherLastname\":\"Monteiro\",\"nationality\":\"PRT\",\"otherInformation\":\"\",\"pan\":\"0000036014662658\",\"requestLocation\":\"CRCiv. Figueira da Foz\",\"socialSecurityNumber\":\"11103478242\",\"username\":\"johndoe\",\"password\":\"12345678\",\"email\":\"johndoe@mail.com\"}"]}'
-# ssh -t ${PEER0_ORG1_IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{\"Args\":[\"person_create\", \"{\"id\":\"4ea88521-031b-4279-9165-9c10e1839001\",\"firstname\":\"John\",\"lastname\":\"Doe\",\"beneficiaryNumber\":\"285191659\",\"birthDate\":\"61985472\",\"cardVersion\":\"006.007.23\",\"country\":\"PRT\",\"documentNumber\":\"09879462 0 ZZ3\",\"documentType\":\"Cartão De Cidadão\",\"emissionDate\":\"61985472\",\"emittingEntity\":\"República Portuguesa\",\"expirationDate\":\"61985472\",\"fatherFirstname\":\"Alberto\",\"fatherLastname\":\"De Andrade Monteiro\",\"fiscalNumber\":\"182692124\",\"gender\":\"M\",\"height\":\"1.81\",\"identityNumber\":\"098794620\",\"motherFirstname\":\"Maria Da Graça De Oliveira Mendes\",\"motherLastname\":\"Monteiro\",\"nationality\":\"PRT\",\"otherInformation\":\"\",\"pan\":\"0000036014662658\",\"requestLocation\":\"CRCiv. Figueira da Foz\",\"socialSecurityNumber\":\"11103478242\",\"username\":\"johndoe\",\"password\":\"12345678\",\"email\":\"johndoe@mail.com\"}\"]}'"
-# invoke chaincode person_create
-ssh -t ${PEER0_ORG1_IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{\"Args\":[\"person_get\", \"4ea88521-031b-4279-9165-9c10e1839001\"]}'"
-# invoke chaincode person_getByUsername
-ssh -t ${PEER0_ORG1_IP} "docker exec cli peer chaincode invoke -C ${CHANNEL} -n ${CHAINCODE_NAME} -c '{\"Args\":[\"person_getByUsername\", \"johndoe\"]}'"
+  echo "bring Production Network Files"
+  ./bringProductionNetworkFiles.sh
+
+  echo "seed ledger with data..."
+  scp seed.env ${PEER0_ORG1_IP}:/tmp
+  scp seedNetwork.sh ${PEER0_ORG1_IP}:/tmp
+  ssh -t ${PEER0_ORG1_IP} sudo chmod a+x /tmp/seedNetwork.sh
+  ssh -t ${PEER0_ORG1_IP} 'cd /tmp; ./seedNetwork.sh'
+fi
+
+# TODO: participant_createWithParameters participant_create don't work via ssh
+echo "now don't forget to:"
+echo "- invoke gov and johndoe, to be tested on sanity-check.sh";
+echo "- run sanity-check.sh, and be patient, until whole network is quick and responsive";
